requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
mysqlclient>=2.1.0
python-decouple>=3.8{
  "name": "John Doe",
  "email": "john.doe@example.com"
}